+++ pyproject.toml
@@ -3,5 +3,5 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "houseprice"
+name = "house_package_saachi"
 version = "0.0.1"
