+++ tests/test_housing_price.py
@@ -3,11 +3,10 @@ import pytest
 
 def test_module():
     try:
-        from src.house_package_saachi import data_ingestion
-        from src.house_package_saachi import model_scoring
-        from src.house_package_saachi import model_training
-        from src.house_package_saachi import utils
-        from src.house_package_saachi import config
-        from src.house_package_saachi import data_preprocessing
+        from house_package_saachi import data_ingestion
+        from house_package_saachi import model_scoring
+        from house_package_saachi import model_training
+        from house_package_saachi import config
+        from house_package_saachi import data_preprocessing
     except ImportError as e:
         assert False, f"Module import failed: {e}"
